backend-api/requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
boto3==1.40.74
botocore==1.40.74
certifi==2025.11.12
cfgv==3.4.0
click==8.3.0
croniter==6.0.0
distlib==0.4.0
fastapi==0.121.2
-e git+ssh://git@github.com/GCUGrayArea/delicious-lotus.git@75fe0ff653d3eceb322933b6ccb3a5120093b025#egg=ffmpeg_backend&subdirectory=ffmpeg-backend
filelock==3.20.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.15
idna==3.11
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.3
PyJWT==2.10.2
nodeenv==1.9.1
platformdirs==4.5.0
pre_commit==4.4.0
psycopg2-binary==2.9.11
pydantic==2.12.4
python-json-logger==2.0.7
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
replicate==1.4.0
rq==2.6.0
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
psutil==6.1.1
openai>=1.0.0
requests>=2.31.0
python-socketio>=5.10.0
